Added "mark entry as complete" functionality
Almost all existing hyperlinks should be working now. The biggest
things left to do right now are interfacing this with the authenticator
such that it has access to the username of who’s logged in, creating
views such that someone can edit their contact info or their classified
ad after it’s been posted, and implementing that truncation feature.