Cleaned up and refactored some code, implemented a few TODOs
Removed some extraneous methods, moved code between forms and views so
that its placement makes more sense. Implemented a feedback feature for
the XP phase, and placed the relevant code at the bottom so that it’s
easy to find and remove when it goes to production. Implemented a
sign-out feature. Implemented feature such that if there are errors on
submission of forms, they will be displayed and the user will not lose
their entered data. Implemented redirects to homepage instead of
re-renders (leaves URL suffix).