Fixed a few bugs found when the site was sent to live on XP
When I changed the display of the classifieds to show the poster’s name
instead of username, I broke a few links that needed the username. I
didn’t find the bug until after pushing the code to H10. Also,  changed
the results pages so that if there are no classifieds to display, it
doesn’t show the table, and instead shows you that none are available.