Mobile css (#4)
* Changes to view messages

* Started with mobile conversion

* Only Banner Blue

Now the banner is only blue and not the whole background.

Table off page so banner doesn’t take full page on mobile (it does on
web)

* Format Table

The mobile-css didn’t have the formatting that made the table look
good.  Looking to play with margins/widths because they still go off
page for mobile

* Add Meta

meta tag used to handle different devices and sizes.  Not fully working
for all devices but some.  NEED TO MESS WITH THE SPECIFICS —> seems to
be a step in the right direction.

* Table made into divs

* Minor conversions and element placement

* Scrollable

Offers scroll and zoom features on mobile version.

* Minor conversions, first attempts at new table

* Update CSS Search/Submit

Fixed margins and padding on both search and submit classified page to
fit better on screen and function better with the footer

* Button Update

Added CSS from old CSS stylesheet to make button look like buttons on
the header

* Updated table

* Table text width updated

Limited the text inside the table to be a certain length/line

* Update

IGNORE

* Banner formatting

* Add CSS for "NO ACTIVE"

CSS added for when there is no active post on page

* Margin Search BUTTON

Added a margin to the submit button on the search page because footer
would cover when screen is compressed

* Updated CSS

updated CSS on some of the profile details

* New submit form

* New contact form and formatting

* Fixed ID

Changed the form ID on the searchPage to be classifieds

* Fixed footer and paragraph padding issues

* Button Margin

All the submit and update buttons have added fixed bottom margin

* Update

Ignore

* Fixed feedback page

* Fixed some formatting issues

* Classified User Table

Changed the view user classifieds to have a table like the front page

* Required fields update, changed contact form

* Fixed cascading issue

* UPDATE

Ignore

* View Public Classified

When clicking on a classified on the home page, the CSS was fixed.
Just used the same class as the other tables.

-WE CAN PLAY AROUND WITH THE LOOK
-LOOKS LIKE THE TABLES WITH THE PROFILE INFORMATION

* Align-Top Description

Put the description of the view_classified table at the top of the cell
and not in the middle of the post

* Update

ignore

* Update

Ignore

* Update viewClassified Length

On the view classified page, the max-length of the text in the
description is modified

* Add MaxLength

Added the max length to all the fields that could be longer than 1 line

* Formatting updates

* Mark as complete link fixed

* Update

Ignore

* Dropdown Button = fixed size

The name button and it’s dropdown content now are the same size so the
dropdown menu is perfect size on all devices

* Aesthetic changes to multiple pages

* Revert "Merge remote-tracking branch 'origin/mobile-css' into mobile-css"

This reverts commit a87baec5346a98dcc773c13e809e702d9b75ffaa, reversing
changes made to 60eabb478ff219bf897f517b894341ef5b029a40.

* Revert "Revert "Merge remote-tracking branch 'origin/mobile-css' into mobile-css""

This reverts commit e8e3148aab3ade4128a9c3513053396e0b6d2b54.

* Revert "Revert "Revert "Merge remote-tracking branch 'origin/mobile-css' into mobile-css"""

This reverts commit 75384bdc828138b94e60909a275c3d4ef1d17bb7.

* Aesthetic changes made to multiple pages, lost previous commit

* Fix Mobile Dropdown content

Dropdown content window matches corner of dropdown button for smaller
size of screen

* Changed cell size

* Tiny Fix to Dropbutton Width

Changed the drop button width a hair because it is was just a little
bit off from the corner of the content

* Button alignment

Aligned buttons with the Bethel Classified title and then aligned the
tables with the new aligned buttons

* Update

Ignore

* Added buttons for sorting

* id change

* adding js for sorting divs

* divs are now sortable, changed repeated ids

* Added buttons for reverse order, new line separator

* Changed buttons to combo box

* Start on conversion to jQuery to get all browsers to work

* Combo Box Update

Changed look of combo box options

* new javascript for using a combobox

* Sort w/ Browsers

Sorts now work on all browsers (chrome, safari, firefox)

* Sorting on viewUserPosts

Now use the combo box to sort on viewUserPosts

* Update

ignore

* Added Comment

update for merge

* Revert "Merge remote-tracking branch 'origin/mobile-css' into mobile-css"

This reverts commit 67c7f24718c937eaa947df3f040614b747498731, reversing
changes made to f7e0bb5da702a720728df678a088de409fccc34d.

* js updates and menu update/change

* Adding image

* Bethel Logo Working

The classified logo not shows up and has margins

* Changes for menu bar (in progress)

* Dropdown fix

Playing around with the dropdown button

* implementing navigation sidebar for mobile

* corrected merge problems

* update

ignore

* Update

ignore

* Update

* Revert "Update"

This reverts commit eb605b3c1559570d8cb322aef6fb5590ec0d7a9f.

* Update

* Changing back to horizontal menu bar

* Test

* More fixing for menu bar

* Update

ignore

* Update

ignore

* New changes after revert, fixing menu

* Fix merge problems

* Delete Crap

* Update

Ignore

* Fix some issues

add background to drop button and delete things that are commented out
and don’t need to happen

* Revert "Fix some issues"

This reverts commit acbcaf0dd75ed3d65eeb12864c902b4aa22bc83f.

* Revert "Merge pull request #1 from betheluniversity/mobile-css"

This reverts commit 25d6c0bdc6ed50c854efc14dec5d6a6bbdcf8be6, reversing
changes made to 06caa1e60af648fc79631f79d8239a95c3b938b1.

Conflicts:
	classifieds/static/stylesheet.css
	classifieds/templates/classifieds-base.html
	classifieds/templates/homepage.html

* Revert "Revert "Merge pull request #1 from betheluniversity/mobile-css""

This reverts commit 509fe6483c5e7ee7cc520d62b89af2bd4e167a8e.

* Fix Footer

* Dropdown content fixed

the drop button has been once again fixed

* Aesthetic changes

* Posts menu and global menu changes

* Cross platform fixes

* Drop button

drop button = good

* Fix min widths

Adjusted min widths so the content wouldn’t be pushed the wrong way

* More cross platform fixes

* Fix min-width

changed the minwidths for the @media tag to avoid the weird jumps that
the menu bar was having

* Touch UP

Fixed multiple little things that now work

* Fixes

Touch ups and menu bar changes

* New Changes

Template names are now consistent, changes made to some labels and
other pieces

* url_for conversion

Converted href paths to url_for statements

* Updated Classifieds

Added descriptions for each page for user, added icon for drop down,
among other simple changes

* Got rid of 'My' in 'My Classified Postings'